feat: realise Controllers
1. realise Services
2. realise Repositories
3. realise Application

Хотелось всё реализовать в точности как в тех задании, поскольку не знаю JavaFX столкнулся с неприятными трудностями под вечер, убил много времени. Нужно это делать на свежую голову, а хочется сделать всё побыстрее, т.к. озвучили: "Чем скорее тем лучше!"

Реализовал некоторые вещи не очень красиво, извините.

+++ src/main/resources/database/init_table.sql
@@ -30,7 +30,7 @@ CREATE TABLE checklines
 (
     line_id      INT8 PRIMARY KEY DEFAULT nextval('global_seq'),
     check_id     INT8                       NOT NULL,
-    goods_ean    INT                        NOT NULL,
+    goods_ean    INT8                       NOT NULL,
     count        INT                        NOT NULL,
     total        FLOAT                      NOT NULL,
     FOREIGN KEY (check_id) REFERENCES checks (check_id) ON DELETE CASCADE
